T: improve 'extract function' refactoring tests
* add mock ui
* use 'ExtractMethod' action instead of 'RsExtractFunctionHandlerAction' direct call
