--- src/main/kotlin/org/rust/lang/refactoring/extractFunction/RsSignatureComponent.kt
@@ -1,18 +1,0 @@
-/*
- * Use of this source code is governed by the MIT license that can be
- * found in the LICENSE file.
- */
-
-package org.rust.lang.refactoring.extractFunction
-
-import com.intellij.openapi.project.Project
-import com.intellij.refactoring.ui.MethodSignatureComponent
-import org.rust.lang.RsFileType
-
-class RsSignatureComponent(
-    signature: String, project: Project
-) : MethodSignatureComponent(signature, project, RsFileType) {
-    private val myFileName = "dummy." + RsFileType.defaultExtension
-
-    override fun getFileName(): String = myFileName
-}
